support for single-value queries
Ignore-this: ca82abb2666fb93246bdd1649410f5d

Philippe Strauss's tutorial exhibited an use for single-value queries
that would, for example, just return the value of a counter. The
present patch implements and try to document this.

A next useful step would be to all aggregate queries such as (SELECT
COUNT(*) FROM foo) to be counted as SQL values. This amount to
special-casing an "aggregate" case of (group .. by ..) where the GROUP
part is a value (instead of a row) and the BY part is empty. The
implementation is, however, not simple (for technical reasons), so
we'll see...

darcs-hash:20120307181852-47e18-d388a21c1914f117b453ce3c2f8e39da2633643e