making Sql.Values output non-results
Ignore-this: ea65c725a69c7008918d1f497b16c83

The initial choice of have Sql.Value.foo value constructors
(eg. Sql.Value.int32 implicitly used by <:value< $int:foo$ >>) return
"immediate value" with a phantom get method allowed to mix them with
results, but is actually problematic when writing some code because
the type with the get method is not unifiable with the type without,
making this example fail:

  let foo = if a then <:value< $bool:true$ >> else <:value< $x$ || $y$ >>

The problem is that the first branch is an immediate value, while the
second is not. Of course, immediate values can be casted back to
non-immediate values, but this is inconvenient to the user.

I decided to forget about this fine point for now, and remove the
"get" phantom method from Sql.Value.foo returns. This might break user
code; if this happens and your code was reasonable, please complain,
and I'll consider reverting this change.

darcs-hash:20120607160443-47e18-115587c44990a893539446f92e1e270c3d226bac