handle reserved SQL keywords by escaping dangerous identifiers at query production time
Ignore-this: 431f0565677074ab2308e06e5d885cf9

This is a big change in the way reserved keywords are handled. After
an initial bug report I added a test at camlp4-time that would warn
the user if one of the identifier used was reserved by the SQL
syntax. I have since then (thanks to a discussion with
Jacques-Pascal Deplaix) found out that it is possible to escape such
"dangerous" identifiers in SQL queries to avoid syntax errors. This
solution is less invasive as it doesn't need to stop or warn the user.

In the hope of minimising the difference between the user-expected and
macaque-produced SQL queries, only the identifiers that are really
reserved are escaped.

darcs-hash:20120920161932-47e18-cc3743f41c4c7a4c5f94ec20811137a053e8539e